[translator] introduce a BaseTranslator base class
It exposes the `BaseTranslator` class that should be used by all concrete
translator classes.

A translator provides a transformation for each available pipeline steps. It
has to be named exactly as the pipeline step name it ought to transform. A
transformation method must return something that matches the
`lib/matchers/OutputStep` type or an array of this type.

The `BaseTranslator` implements the `StepMatcher<OutputType>` type so
typescript will make sure each transform method is implemented on this
class. This class should not provide any concrete transform
implementation. Instead, it should use the `@unsupported` decorator that
will wrap the transformation method to make it raise a
`StepNotSupported` exception. It will also set an hidden
`__vqb_step_supported__` property that will be used by the
`supportXXX()` methods to detecte dynamically which steps are supported
or not by a given translator instance.

The `@unsupported` decorator option was chosen over the use of the
`abstract` keyword because it would forbid intanciation of
`BaseTranslator` subclasses unless they provide an implementation for
all possible steps.