[translators] expose registerTranslator and use it
Instead of exposing a random `translators` in a translator module,
expose the generic `registerTranslator` function and make it the
official API for a module to register a specific translator.

Since the `mongo36` translator should be bundled by default in our
library, we nevertheless register it manually inside the library to that
the client code doesn't have to do on its side.

On the way, add an `availableTranslators` function in the `translators`
module that returns the list of all available translators.