[translator] force callback type cast
Since we chose the 'discriminated union' approach, when adding a new
step type, we had to add a new `case` condition to the giant `switch`
statement of the `translate` method. The idea was to let TypeScript check
at compile time that we didn't forget to handle a step type but in our
case, this is painful because the `case` code is always the same _and_
because the compile-time check is already enforced by the `implements
StepMatcher` condition on the `BaseTranslator` class.

I could not find a typescript way to remove the need of `case` so I just
decided to cheat and expclitly cast to a valid callback type.

Closes #98.