Use our own type declaration file for mathjs
We decide not to use the standard one because of the lack of precision on
node interfaces. This type declaration concentrates purely on what we need
therefore the vast majority of the `mathjs` API is not covered here.

We'll probably end up adding a PR on the original project.

Related to #67