feature :: standalone example (#125)
Introduce a new standalone playground script. It consists of a simple [express](https://expressjs.com/) server that essentially exposes 3 routes:

- `/` to serve the VQB components (mostly `ResizablePanel` with a `Pipeline` on the left and a `DataViewer` on the right),

- `/query` to perform a mongo query on a real database

- `/collections` to list the mongo database collection

The database used can be configured through a config file or command line options. New targets (`playground`, `playground:build-app` and `playground:build-server`) were added to the `package.json` file. The transformation pipeline is hardcoded for now but the whole purpose of the VQB is to provide a tool to edit it ;-)

Refer to the `README.md` for more detailed information on how to do this.

Closes #34