[lib] expose filterOutDomain
`filterOutDomain` extracts the requested domain from the first step and
return the rest of the pipeline unchanged. This is useful in the GUI
where the domain steps is handled differently from the rest of the
pipeline. It's also useful in backends such as mongo where the first
step corresponds to the collection that has to be queried.