extract button definition in a component
Introduce the `ButtonDef` type with 2 properties `icon` and `label`. We will
probably add `id` or `category` later on.

The `ActionToolbarButton` component is fed with a `ButtonDef` instance: it uses
the `icon` property to fetch the corresponding font-awesome icon and `label` for
the button label.

`Search` and `Column` buttons are left hardcoded for now.