[karma] single test-suite entrypoint
This way, we don't have to rely on `rollup-plugin-multi-entry` and
separate `karma` and `rollup` process. Everything is handled through
rollup preprocessing specified in karma configuration.

There's one tiny hack to make the transpiled suite loadable in karma: we
have to enforce `.ts` mimetypes so that chrome accepts to load it.