Stop listening to setPipeline mutaton in backend-plugin
Except at init, this mutation is always combined with a selectStep mutation which is
already watched, while selectStep mutation can be fired without setPipeline. We want
to execute the query only once when the selectStep mutation is fired.

At init, we already watch setDomain mutation so still no need for setPipeline