[storybook] fix rollup config
rollup bundling fixes the `__dirname` value to the entry point's dirname
rather than the actual module one. This could probably be fixed with a specific
plugin or by playing with the plugin order but I chose to implement a custom
`packageDir()` function that simply goes up in the directory hierarchy until
reaching the directory where `package.json` lies.