[stepforms] first draft of StepForm baseclass
- introduce the `$$super` method

The `$$super` property will return a javascript proxy that should bind
the current `this` instance on the method implentation found on the baseclass.

This is required because otherwise `@Component` will flatten the prototype chain
and we won't be able to access the base class anymore. Of course, this relies on
the idea that `$$super` must not be overridden on a subclass. Furthermore, current
implementation assumes that we won't go deeper than 1 level of subclassing.