[steps] make step type definitons writable
Step type definitions have no good reason to be readonly. In fact,
it is actually a pain to have steps being readonly in step forms since
we can't use the step types directly on a data attribute without wrapping
them with a specific `Writable` type that removes the `readonly` modiifier.

What we actually want is to make steps readonly during the translation process,
so just declare transform step function parameters as such and everything is
fine.