[translator] fix handling of empty mongo pipelines
Before this commit, `simplifyMongoPipeline` would output a `[null]`
mongo pipeline if fed with an empty one. This is of course incorrect and
is caused by the following statements at the beginning of the function:

```typescript
  const outputSteps: MongoStep[] = [];
  let lastStep: MongoStep = mongoSteps[0];
  outputSteps.push(lastStep);
```

This commit doesn't change this specific logic but instead explicitly
tests for an empty input pipeline and return another empty one in that
case.