[variables] provide basic interpolator class
The interpolator class is responsible for "interpolating" variables
in each step. It is passed an "interpolator" functino (e.g. `_.template`)
and a context scope (i.e. a list of variable definitions) and it provides
an `interpolate` method that iterates on each step and tries to interpolate
step fields that can contain "variable" slots.

related to #320