[playground] make mongo return types
Before that change, we could not interpret correctly a date column
as it was just returned as JSON string from Mongo.
We could have tried to parse any string column into a date but it
is not satisfying as in a lot of case the user wants to keep a
string format even if it is a date string.

So we rather implemented a solution that relies on the data type
returned from Mongo. Instead of looping through the original
dataset documents and guessing every value datatype, we now loop
through "meta" documents that has explicit information about the
data type of every value rather than the value itself.