[concatenate] fix bad state sync on edition
Before this commit, when a concatenate step was edited, its "column"
fields were reset with the current selected column in the dataset
viewer.

This was because the `value` property sent by the `List` component to
the `ColumnPicker` which contains the edited step's column names was not
taken into account by the `ColumnPicker`. Instead, it was introspecting
an `initialColumn` property that was never passed in this case and
actually never used elsewhere…

This commit gets rid of the `initialColumn` property and uses `value`
instead.

Closes #366