[tests] create required store for text-button related tests
The action-toolbar-button tests will need to access the `selectedColumns` getter
of the store to determine whether or not a form has to be shown: remember that
if a column is selected, steps such as `lowercase` will be directly inserted in
the pipeline without any form popping out. Therefore, we need to have a valid
and wired Vuex store beforehand.

This will get rid of the `Cannot read property '_modulesNamespaceMap' of undefined` error