[steps] Add append step implementation
This step allows to append several datasets, referenced as
pipelines names in the application store.

Implement a simple dereference function rather than a class method
Remove the PipelineDereferenced class introduced in the previous
commit and instead of using the `dereference` class method from
this class, implement a simple `dereferencePipelines` fucntion
directly in the backend plugin.