keep history of error messages
This commit introduces two new mutations:

- `logBackendError` to add a backend error to the logbook
- `resetErrorMessages` to empty the logbook

The former `setBackendError` mutation is removed.

Therefore, rather than having a single error message in the store, keep
track of all of them until the `resetBackendErrors` is called.

This is needed because otherwise the backend-plugin would reset a
previous error once a pipeline execution was successful. Since we
sometimes peform several pipeline executions in a row (e.g. if 2
mutations trigger a pipeline execution consecutively).