add a `defaults` parameter to {open,create}FormStep
and pass it through the `actionClicked` event.

The `defaults` parameter will allow steps to be instantiated and pre-initialized
with custom values. This will be useful, for instance, for the upcoming
"extract date-property from date" steps where we actually have a single step model
but we want a different menu item for each possible date property. With this
`defaults`, we will be able to initialize the step model with the corresponding
operation (e.g. `minutes`, `day`, `year`)