Map domains to collection with a translator method
Previously, we expected a map of domains to collections for the mongo
translator. This supposed that we already know all the available
domains and their correspondong collections. This is however not always
the case, e.g. if domains needs asynchronous loading.
I decided to replace it by a function, which cover the previous
case easily, and provide more possibilities.