[playground] change default dataset
Add a `datastore` directory and a specific `sales.csv` file which is a
bit more stuffed and useful than the previous default one. Besides, this
very file will also be available on a free mongo atlas endpoint that
we'll use later to build a free online demo for the playground. Having a
useful shared default case / default pipeline will help.