setup heroku continuous deployment
Since the playground is quite standalone, deploying on heroku is
just a matter of having setup a mongodb public endpoint (which I did)
with some default data (e.g. `playground/datastore/sales.csv`).

Heroku was setup and configured to query this endpoint so there is nothing
specific to do (beyond github plumbing) for the continuous deployment to
work.

To avoid the need of a `Procfile` or a `heroku.yml` file, use
conventional scripts entry points such as `build` and `start` that
many platforms (including heroku) will be able to understand.