reset pagination after some mutations
When navigating in or modifying the pipeline, the pagination needs
to be reset to avoid inconsistencies.

e.g consider a pipeline with 4 steps and a dataset on page 3. If the last step
is a filter step and the user removes it, the page `3` does not have the same
meaning. The only consistent behaviour is to go back to page 1 with the new
dataset.