refactor: upgrade vue-class-component and rework BaseStep
The @StepFormComponent was relying on a behavior that has changed in
vue-class-component/vue-loader: the config passed to the decorator does not have the rendering function
in it anymore, which means it never gets included in the registry, and
therefor cannot be instantiated this way.
I decided to keep things simple, and revert to a more usual / less magic
way of instantiating the form steps.