feat(front): table numbers improvements (#1092)
* feat(front): use fixed width numbers in table

* Align numeric cells to right

* refactor(front): cell is not more functional

Functional components deal with some problems regarding the classes
applied by their parent. As they are gonna disappear in vue3, and thay
they do not have such a performance improvement, I suggest we remove the
functional part of the cell. Moreover, this make it possible to use
computed, so each time the cell is re-rendered, these 2 will not be
re-computed again

* test(front): adapt DataViewer tests (no more functional)

And add the numerical test case

* test(front): move cell selection logic to data viewer