fix(mongo_translator/evolution): Replace manual date substraction with mongo's  operator
There is a bug in the evolution step: when using $dateFromParts, we only keep a hourly granularity, meaning
that datetimes with a sub-hourly precision did never match our predicted offset.

This will not be backported to the TypeScript translator, as it must support mongo<5, and will eventually be
replaced by the Python implementation.

Signed-off-by: Luka Peschke <luka.peschke@toucantoco.com>