fix(Formula AST): Be more permissive on column names TCTC-3866
Allow column names to contain operators and extra whitespaces.

An implementation change is required for that: We now store the start and end position of bracket
delimiters and consider everything in-between to be the column name (meaning we don't strip whitespace
at the start and end anymore, nor do we check the validity of the tokens in the column name).

Also allow string literals to contain quotes

TCTC-3866

Signed-off-by: Luka Peschke <luka.peschke@toucantoco.com>