fix(backend/pandas/addmissingdates): Ensure we always insert pd.Timestamp objects
Before this change, in case we had timezone-aware series, inserting the start of a preiod (naive) would result
in an integer (representing a unix timestamp in nanoseconds). The resulting series would end up being a mix of
tz-aware timestamp objects and integer (for naive inserted dates).

This ensure that inserted objects are properly localized if required.

closes TCTC-4429