feat(sql): change how first and last aggregation are translated
There is no easy FIRST and LAST aggregation functions in SQL.
The current approach with window functions is very inefficient, and causes queries to time out.
For now, we suggest a quick fix: to use min and max for first and last respectively.
This is way faster is many situations.