chore(ui): remove engine (#1924)
The engine field is used to specify which node version should be used
to _use_ the package, not to _build_ it.
Once built, the engine required is actually a browser :joy:

I removed this field, so that dependent packages could choose which node
version they should use. They will use the pre-built js files
from weaverbird anyway, without the need to build them again.