fix(server/pypika): fix _id method (#2097)
The _id method appears to be flaky: it relies on the id() built-in, which must return a unique identifier for an
object. In CPython, this is the object's address.

This causes an issue in the `append` method of the PyPika translator when several pipelines are appended:
an instance of the translator class is created everytime we need to process a pipeline. We assume the id of
every instance to be unique to generate unique human-readable step names in CTEs.

However, when multiple pipelines are appended, an instance might be garbage-collected between two iterations
of the for-loop, which can resulting in the next instance (or the one after that) to be allocated in the
address as the garbage-collected one. This results in both instances having the same id, and thus generating
the same step names.

This issue is fixed by generating a unique id in the classe's __init__ method, and rely on this id rather than the
object's pythonic id when generating step names.

Signed-off-by: Luka Peschke <luka.peschke@toucantoco.com>